Fixed the registration crashing.
The error was due to a function being called via the microsyntax which caused the app to appear to have a memory leak of sorts. Inside the chrome dev tools it shows the memory increase drasitcally.

To fix this I have added two new fields to the RegData class, hasRelatedVehicle and relatedVehicle, which gets set when the data is recieved from the event emitter.